REF: split pvpositioner into put-complete and non-put complete version
