+++ ophyd/controls/pv_positioner.py
@@ -99,12 +99,18 @@ class PVPositioner(Positioner):
             raise ValueError('A setpoint or a readback must be specified')
 
         if self.done is None and not self.put_complete:
-            msg = ('Positioner %s is mis-configured. A "done" Signal must be '
-                   'provided or put_complete must be True.'.format(self.name))
+            msg = ('PVPositioner %s is mis-configured. A "done" Signal must be '
+                   'provided or use PVPositionerPC (which uses put completion '
+                   'to determine when motion has completed).'
+                   ''.format(self.name))
             raise ValueError(msg)
 
         if self.done is not None:
             self.done.subscribe(self._move_changed)
+
+    @property
+    def put_complete(self):
+        return isinstance(self, PVPositionerPC)
 
     def check_value(self, pos):
         '''Check that the position is within the soft limits'''
@@ -133,79 +139,26 @@ class PVPositioner(Positioner):
         else:
             return self._moving
 
-    def _move_wait_pc(self, position, **kwargs):
-        '''*put complete* Move and wait until motion has completed'''
-        has_done = self.done is not None
-        if not has_done:
-            moving_val = 1 - self.done_value
-            self._move_changed(value=self.done_value)
-            self._move_changed(value=moving_val)
-
-        timeout = kwargs.pop('timeout', self._timeout)
-        if timeout <= 0.0:
-            # TODO pyepics timeout of 0 and None don't mean infinite wait?
-            timeout = 1e6
-
-        if self.actuate is None:
-            self.setpoint.put(position, wait=True, timeout=timeout)
-        else:
-            self.setpoint.put(position, wait=False)
-            self.actuate.put(self.actuate_value, wait=True, timeout=timeout)
-
-        if has_done:
-            time.sleep(self.settle_time)
-        else:
-            self._move_changed(value=self.done_value)
-
-        if self._started_moving and not self._moving:
-            self._done_moving(timestamp=self.setpoint.timestamp)
-        elif self._started_moving and self._moving:
-            # TODO better exceptions
-            raise TimeoutError('Failed to move %s to %s '
-                               '(put complete done, still moving)' %
-                               (self.name, position))
-        else:
-            raise TimeoutError('Failed to move %s to %s '
-                               '(no motion, put complete)' %
-                               (self.name, position))
-
     def _move_wait(self, position, **kwargs):
         '''Move and wait until motion has completed'''
         self._started_moving = False
 
-        if self.put_complete:
-            self._move_wait_pc(position, **kwargs)
-        else:
-            self.setpoint.put(position, wait=True)
-            logger.debug('Setpoint set: %s = %s',
-                         self.setpoint.setpoint_pvname, position)
+        self.setpoint.put(position, wait=True)
+        logger.debug('Setpoint set: %s = %s',
+                     self.setpoint.setpoint_pvname, position)
 
-            if self.actuate is not None:
-                self.actuate.put(self.actuate_value, wait=True)
-                logger.debug('Actuating: %s = %s',
-                             self.actuate.setpoint_pvname, self.actuate_value)
+        if self.actuate is not None:
+            self.actuate.put(self.actuate_value, wait=True)
+            logger.debug('Actuating: %s = %s',
+                         self.actuate.setpoint_pvname, self.actuate_value)
 
     def _move_async(self, position, **kwargs):
         '''Move and do not wait until motion is complete (asynchronous)'''
-        self._started_moving = False
-
-        def done_moving(**kwargs):
-            if self.put_complete:
-                logger.debug('%s async motion done', self.name)
-                self._done_moving()
-
-        if self.done is None and self.put_complete:
-            # No done signal, so we rely on put completion
-            moving_val = 1 - self.done_value
-            self._move_changed(value=moving_val)
-
         if self.actuate is not None:
             self.setpoint.put(position, wait=False)
-            self.actuate.put(self.actuate_value, wait=False,
-                             callback=done_moving)
+            self.actuate.put(self.actuate_value, wait=False)
         else:
-            self.setpoint.put(position, wait=False,
-                              callback=done_moving)
+            self.setpoint.put(position, wait=False)
 
     def move(self, position, wait=True, **kwargs):
         try:
@@ -215,6 +168,7 @@ class PVPositioner(Positioner):
             else:
                 # Setup the async retval first
                 ret = super().move(position, wait=False, **kwargs)
+                self._started_moving = False
                 self._move_async(position, **kwargs)
                 return ret
         except KeyboardInterrupt:
@@ -270,3 +224,61 @@ class PVPositioner(Positioner):
                 'limits={0._limits!r}'.format(self)
                 ]
         return self._get_repr(repr)
+
+
+class PVPositionerPC(PVPositioner):
+    def _move_wait(self, position, **kwargs):
+        '''Move and wait until motion has completed'''
+        self._started_moving = False
+        has_done = self.done is not None
+        if not has_done:
+            moving_val = 1 - self.done_value
+            self._move_changed(value=self.done_value)
+            self._move_changed(value=moving_val)
+
+        timeout = kwargs.pop('timeout', self._timeout)
+        if timeout <= 0.0:
+            # TODO pyepics timeout of 0 and None don't mean infinite wait?
+            timeout = 1e6
+
+        if self.actuate is None:
+            self.setpoint.put(position, wait=True, timeout=timeout)
+        else:
+            self.setpoint.put(position, wait=False)
+            self.actuate.put(self.actuate_value, wait=True, timeout=timeout)
+
+        if has_done:
+            time.sleep(self.settle_time)
+        else:
+            self._move_changed(value=self.done_value)
+
+        if self._started_moving and not self._moving:
+            self._done_moving(timestamp=self.setpoint.timestamp)
+        elif self._started_moving and self._moving:
+            # TODO better exceptions
+            raise TimeoutError('Failed to move %s to %s '
+                               '(put complete done, still moving)' %
+                               (self.name, position))
+        else:
+            raise TimeoutError('Failed to move %s to %s '
+                               '(no motion, put complete)' %
+                               (self.name, position))
+
+    def _move_async(self, position, **kwargs):
+        '''Move and do not wait until motion is complete (asynchronous)'''
+        def done_moving(**kwargs):
+            logger.debug('%s async motion done', self.name)
+            self._done_moving()
+
+        if self.done is None:
+            # No done signal, so we rely on put completion
+            moving_val = 1 - self.done_value
+            self._move_changed(value=moving_val)
+
+        if self.actuate is not None:
+            self.setpoint.put(position, wait=False)
+            self.actuate.put(self.actuate_value, wait=False,
+                             callback=done_moving)
+        else:
+            self.setpoint.put(position, wait=False,
+                              callback=done_moving)
